move script to onnx folder
